+++ src/malgtk_malitem.c
@@ -15,9 +15,7 @@
  *  along with mal-gtk.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <stdio.h>
 #include <string.h>
-#include <glib.h>
 #include "malgtk_malitem.h"
 #include "malgtk_date.h"
 #include "malgtk_enumtypes.h"
@@ -736,7 +734,7 @@ static const struct {
 static gint
 find_field_for_name(const xmlChar *name)
 {
-    for (guint i = 0; i < G_N_ELEMENTS(field_map); ++i)
+    for (size_t i = 0; i < G_N_ELEMENTS(field_map); ++i)
     {
         if (xmlStrEqual(field_map[i].xml_name, name)) {
             return i;
